+++ lib/src/goods/detail/guide.dart
