+++ lib/src/goods/detail/index.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'top.dart'; // top Image wrap
 import 'title.dart'; // title wrap
-import 'guilde.dart';
+import 'guide.dart';
 import 'salePoints.dart';
 import 'routePoint.dart';
 import 'map.dart';
@@ -90,19 +90,19 @@ class HbcGoodsIndex extends StatelessWidget {
       return '';
     }
 
-    Map route1 = {
+    final Map route1 = {
       'imgUrl': 'https://fr-static.huangbaoche.com/20180305/route-info1.66769120966ac3f8.png',
       'title': '${dayCount}天,(约${totalDistance}公里)',
       'subTitle': '${routeCityDesc}'
     };
-    Map route2 = {
+    final Map route2 = {
       'imgUrl': 'https://fr-static.huangbaoche.com/20180305/route-info2.3a7c8c14c30f8a67.png',
       'title': '上车时间可协调更改，建议${boardingTimeStart}${boardingTimeEnd != null
           ? '~${boardingTimeEnd}'
           : ''}',
       'subTitle': '酒店门口／机场／指定地点接送'
     };
-    Map route3 = {
+    final Map route3 = {
       'imgUrl': 'https://fr-static.huangbaoche.com/20180305/route-info3.ac42ea4631d85fd8.png',
       'title': '中文服务，可接待(1${maxCapOfPerson != 0 ? '~${maxCapOfPerson}' : ''}人)',
       'subTitle': _getRout3SubTitle(),
